+++ src/main/java/com/mentalhealthforum/mentalhealthforum_backend/dto/userProfileAndIdentity/auth/OtpResult.java
@@ -1,3 +1,3 @@
 package com.mentalhealthforum.mentalhealthforum_backend.dto.userProfileAndIdentity.auth;
 
-public record OtpResult(String code, int expiryMinutes) {}
+public record OtpResult(String code, long expiryMinutes) {}
